+++ src/main/java/de/samply/qualityreportgenerator/exporter/RequestResponseEntity.java
@@ -1,0 +1,7 @@
+package de.samply.qualityreportgenerator.exporter;
+
+public record RequestResponseEntity(
+    String responseUrl
+) {
+
+}