+++ CHANGELOG.md
@@ -10,3 +10,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Dockerignore
 - build.yml
 - Dockerfile
+- Exporter Client